Running on real hardware! Starting to implement the pmm

+++ kernel/stdlib/util.h
@@ -2,8 +2,9 @@
 #define UTIL_H
 
 #include <stddef.h>
+#include <stdint.h>
 
-char* itoa(size_t value, char *str, size_t base);
+char* itoa(uint64_t value, char *str, size_t base);
 void* memset(void *dest, int val, size_t len);
 void* memcpy(void *dest, const void *src, size_t len);
 void* kmalloc(size_t size);
